Update menu.py
I found an inappropriate DropDownMenu display in a situation that had rendered items in the negative values region of the y-axis (below the visible area of the window). With this failure, the respective items were hidden and inaccessible by the ScroolView roll. My contribution is just to alert to find a better solution to this issue.